Replace clap with structopt CLI
