+++ src/main.rs
@@ -1,269 +1,13 @@
+mod cli;
 mod config;
 mod repository;
 
-use anyhow::{anyhow, Context, Result};
-use clap::{crate_authors, crate_description, crate_name, crate_version, App, Arg, SubCommand};
-use config::Config;
-use directories::ProjectDirs;
-use indicatif::{ProgressBar, ProgressStyle};
-use prettytable::{cell, format, row, Table};
-use repository::Repository;
-use std::collections::BTreeMap;
-use std::path::Path;
-use std::str::FromStr;
-use std::sync::mpsc::channel;
-use threadpool::ThreadPool;
+use cli::CLI;
+
+use anyhow::Result;
+use structopt::StructOpt;
 
 fn main() -> Result<()> {
-    // Get default config path
-    let app_name = crate_name!();
-    let project_dirs = ProjectDirs::from("com", app_name, app_name)
-        .ok_or_else(|| anyhow!("could not retrieve home directory from system"))?;
-    let default_config_path = project_dirs.config_dir().join("config.toml");
-
-    // Create clap app
-    let matches = App::new(app_name)
-        .version(crate_version!())
-        .author(crate_authors!("\n"))
-        .about(crate_description!())
-        .arg(
-            Arg::with_name("config")
-                .short("c")
-                .long("config")
-                .value_name("FILE")
-                .default_value_os(&default_config_path.as_os_str())
-                .help("Sets a custom config file")
-                .takes_value(true),
-        )
-        .arg(
-            Arg::with_name("nofetch")
-                .short("F")
-                .long("no-fetch")
-                .help("Do not fetch"),
-        )
-        .arg(
-            Arg::with_name("workers")
-                .long("workers")
-                .value_name("NUM_WORKERS")
-                .default_value("4")
-                .help("Number of workers")
-                .takes_value(true),
-        )
-        .subcommand(
-            SubCommand::with_name("add")
-                .about("Add new repositories")
-                .arg(
-                    Arg::with_name("path")
-                        .value_name("PATH")
-                        .multiple(true)
-                        .help("Paths to the repositories to add")
-                        .required(true),
-                ),
-        )
-        .subcommand(
-            SubCommand::with_name("remove")
-                .about("Remove repositories")
-                .arg(
-                    Arg::with_name("name")
-                        .value_name("NAME")
-                        .multiple(true)
-                        .help("Names of the repositories to remove")
-                        .required(true),
-                ),
-        )
-        .subcommand(
-            SubCommand::with_name("rename")
-                .about("Rename repository")
-                .arg(
-                    Arg::with_name("name")
-                        .value_name("NAME")
-                        .help("Current name of the repository")
-                        .required(true),
-                )
-                .arg(
-                    Arg::with_name("new_name")
-                        .value_name("NEW_NAME")
-                        .help("New name of the repository")
-                        .required(true),
-                ),
-        )
-        .subcommand(
-            SubCommand::with_name("path")
-                .about("Get repository's path")
-                .arg(
-                    Arg::with_name("name")
-                        .value_name("NAME")
-                        .help("Name of the repository")
-                        .required(true),
-                ),
-        )
-        .get_matches();
-
-    // Create config directory if it doesn't exist
-    let config_dir = project_dirs.config_dir();
-    if !config_dir.is_dir() {
-        std::fs::create_dir(config_dir).context("failed to create config directory")?;
-    }
-
-    // Load config from file or create new one
-    let config_path = Path::new(matches.value_of("config").unwrap());
-    let mut config: Config = if config_path.is_file() {
-        Config::from_path(config_path).context("failed to load config")?
-    } else {
-        Config::new()
-    };
-
-    if let Some(subcommand_name) = matches.subcommand_name() {
-        // Track config modification
-        let mut modified = false;
-
-        match (subcommand_name, matches.subcommand_matches(subcommand_name)) {
-            ("add", Some(matches)) => {
-                for path in matches.values_of("path").unwrap() {
-                    let path = Path::new(path);
-                    config.add_repository(path)?;
-                    modified = true;
-                }
-            }
-            ("remove", Some(matches)) => {
-                for name in matches.values_of("name").unwrap() {
-                    if config.remove_repository_by_name(name) {
-                        modified = true;
-                    }
-                }
-            }
-            ("rename", Some(matches)) => {
-                let name = matches.value_of("name").unwrap();
-                let new_name = matches.value_of("new_name").unwrap();
-                config.rename_repository(name, new_name)?;
-                modified = true;
-            }
-            ("path", Some(matches)) => {
-                let name = matches.value_of("name").unwrap();
-                let path = config
-                    .repositories()
-                    .get(name)
-                    .context("name does not exist")?;
-                println!("{:?}", path);
-            }
-            (invalid_subcommand, _) => {
-                eprintln!("Invalid subcommand '{}'", invalid_subcommand);
-            }
-        }
-
-        // Save config
-        if modified {
-            config.save(config_path).context("failed to save config")?;
-        }
-
-        // Exit without further processing
-        std::process::exit(0);
-    }
-
-    // Attempt to open repositories
-    let mut repositories = Vec::with_capacity(config.repositories().len());
-    for (name, path) in config.repositories() {
-        match Repository::open(name, path) {
-            Ok(repository) => repositories.push(repository),
-            Err(e) => eprintln!("Could not open '{}': {}", name, e),
-        }
-    }
-
-    // Create thread pool
-    let num_workers = usize::from_str(matches.value_of("workers").unwrap())
-        .context("invalid number of workers")?;
-    let pool = ThreadPool::new(num_workers);
-    let (tx, rx) = channel();
-    let num_jobs = repositories.len();
-
-    // Create progress bar
-    let pb = ProgressBar::new(num_jobs as u64);
-    pb.set_style(
-        ProgressStyle::default_bar()
-            .template("{prefix} [{bar:60}] {pos}/{len}: {msg}")
-            .progress_chars("=> "),
-    );
-    pb.set_prefix("Processing...");
-
-    // To fetch or not to fetch
-    let do_fetch = !matches.is_present("nofetch");
-
-    // Process repositories on thread pool
-    for repository in repositories.into_iter() {
-        let mut repository = repository;
-        let tx = tx.clone();
-        let pb = pb.clone();
-
-        pool.execute(move || {
-            // Attempt to fetch from repository
-            if do_fetch {
-                let _ = repository.fetch();
-            }
-            // Compute status now since it can be slow
-            let _ = repository.compute_status();
-
-            // Update progress bar
-            pb.set_message(repository.name());
-            pb.inc(1);
-
-            tx.send(repository).unwrap();
-        });
-    }
-
-    // Join threads and collect data in a sorted map
-    let sorted_map = rx
-        .iter()
-        .take(num_jobs)
-        .fold(BTreeMap::new(), |mut map, repository| {
-            map.insert(repository.name().to_string(), repository);
-            map
-        });
-
-    // Clear progress bar
-    pb.finish_and_clear();
-
-    // Create table
-    let mut table = Table::new();
-
-    // Format table
-    let format = format::FormatBuilder::new()
-        .column_separator(' ')
-        .borders(' ')
-        .padding(0, 3)
-        .build();
-    table.set_format(format);
-
-    // Add rows to table
-    for (name, repository) in sorted_map.iter() {
-        // Get status
-        let status = if let Some(status) = repository.status() {
-            status.to_string()
-        } else {
-            String::new()
-        };
-        // Get distance between local and upstream
-        let distance = if let Some(distance) = repository.distance() {
-            distance.to_string()
-        } else {
-            String::new()
-        };
-        table.add_row(row![
-            name,
-            status,
-            repository.branch_name().unwrap_or_default().to_string(),
-            distance,
-            repository.remote_name().unwrap_or_default().to_string(),
-            repository
-                .commit_summary()
-                .unwrap_or_default()
-                .chars()
-                .take(50)
-                .collect::<String>()
-        ]);
-    }
-
-    // Display table
-    table.printstd();
-
-    Ok(())
+    let cli = CLI::from_args();
+    cli.run()
 }
